mini-services/free-video-gen/requirements.txt
# Free Video Generator - Dependencies
# AI Video Generation from free providers

# Web Automation
playwright>=1.40.0
playwright-stealth>=1.0.6

# HTTP & API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
aiohttp>=3.9.0

# Video Processing (FFmpeg wrapper)
ffmpeg-python>=0.2.0

# Audio Processing
pydub>=0.25.1
gtts>=2.4.0

# Queue & Async
asyncio-pool>=0.6.0
aiofiles>=23.2.1

# Database & Cache
aiosqlite>=0.19.0

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0
PyYAML>=6.0.1
python-dotenv>=1.0.0

# Logging
loguru>=0.7.2
rich>=13.7.0

# Utilities
python-multipart>=0.0.6
orjson>=3.9.10
pillow>=10.2.0

# Browser automation helpers
undetected-chromedriver>=3.5.4
selenium>=4.16.0

# Progress bars
tqdm>=4.66.1

# Optional: Local LLM for scene splitting
# llama-cpp-python>=0.2.20
